feat: send pg_notify on team_members update
* on teams delete cascade it to team_members
* add TG_OP to the payload for `table_activity` notifications
* for notifications on `team_members` table, send both team_id and person_id instead of just the primary key like in the other tables

**TODO:**
When a team is deleted, we send pg notifications for all the members associated with the team
and another notification for the deletion of the team. For our use cases, we just need one notification _(i.e. of the team's deletion)_, so we can just exec one `DeleteRole()` call instead of `DeleteRoleForPerson()` for each members.