Rename RLS migration scripts to execute last (#1599)
we need to conditionally run RLS related scripts depending on

whether RLS is enabled or not
or, whether the system is running on postgres 14 or below (which doesn't support SECURITY INVOKER on views)
With RLS scripts being named 034_rls_enable.sql it would run before some views that would be created on scripts that come later on.
This change ensures that the RLS scripts are run at the end.